exchange Landingpage Map for Screenshot, move Map to WalkDetail, change Directory Structure for Components
