+++ Frontend/src/components/navigation/NavigationButton.vue
