+++ Frontend/src/views/LandingPage.vue
@@ -1,5 +1,5 @@
 <template>
-  <v-app id="inspire">
+  <v-app id="photowalkapp">
     <v-app-bar
         app class="gradient"
         shrink-on-scroll
@@ -15,10 +15,15 @@
       <v-spacer></v-spacer>
       <v-row>
         <v-col cols="12" md="6">
-            <PhotowalkMap/>
+            <v-card class="container card ma-4">
+              <v-img
+                  :src="`/images/map_landingpage_placeholder.JPG`"
+                  :lazy-src="`images/gradient.png`"
+                  class="rounded-lg elevation-1 grid-image"></v-img>
+            </v-card>
         </v-col>
         <v-col cols="12" md="6">
-          <div class="px-2">
+          <div class="mx-2 card">
           <ImageGrid/>
           </div>
           <LoginDialog/>
@@ -37,18 +42,12 @@
 
 <script>
 import LoginDialog from "../components/usermanagement/LoginDialog";
-import PhotowalkMap from "@/components/PhotowalkMap";
 import ImageGrid from "@/components/ImageGrid";
 import Impressum from "@/components/legal/Impressum";
 import DGSVO from "@/components/legal/DGSVO";
 export default {
   name: 'LandingPage',
-  components: {DGSVO, Impressum, ImageGrid, PhotowalkMap, LoginDialog},
-  data () {
-    return{
-      impressum : false,
-    };
-},
+  components: {DGSVO, Impressum, ImageGrid, LoginDialog}
 }
 </script>
 
@@ -57,8 +56,13 @@ export default {
   background: linear-gradient(90deg, #00BCD4 0%, #E91E63 100%);
   color: white;
 }
-
 .headline{
   justify-content: center;
+}
+.container {
+  padding: 10px;
+}
+.card{
+  margin-top: 15px;
 }
 </style>
