+++ Frontend/src/views/WalksDetail.vue
@@ -1,12 +1,17 @@
 <template>
-  <!--vuetify div-->
-  <v-container>
-    <p>WALK DETAIL 1-3</p>
-  </v-container>
+  <v-main>
+    <v-row>
+      <v-col cols="12" md="6">
+        <PhotowalkMap/>
+      </v-col>
+    </v-row>
+  </v-main>
 </template>
 
 <script>
+import PhotowalkMap from "@/components/PhotowalkMap";
 export default {
   name: 'WalksDetail',
+  components: {PhotowalkMap},
 }
 </script>
